Add Babel/Webpack, set up package.json and webpack config files. Modularize and move files around. Begin rewriting code in ES6..
