+++ package.json
@@ -1,0 +1,46 @@
+{
+  "name": "glasses-program",
+  "version": "1.2.0",
+  "description": "Glasses inventory and search program used by Global Medical Brigades chapters working in Honduras, Panama, Nicaragua",
+  "main": "src/app.js",
+  "scripts": {
+    "dev": "node_modules/.bin/webpack --config webpack.dev.js",
+    "build": "node_modules/.bin/webpack --config webpack.build.js",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/pll33/glasses-program.git"
+  },
+  "keywords": [
+    "inventory",
+    "angular1",
+    "pouchdb",
+    "papaparse",
+    "csvparse",
+    "glasses",
+    "eyeglasses"
+  ],
+  "author": "Peter Liu",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/pll33/glasses-program/issues"
+  },
+  "homepage": "https://github.com/pll33/glasses-program#readme",
+  "private": true,
+  "devDependencies": {
+    "babel-core": "^6.26.0",
+    "babel-loader": "^7.1.2",
+    "babel-polyfill": "^6.26.0",
+    "babel-preset-env": "^1.6.1",
+    "clean-webpack-plugin": "^0.1.17",
+    "copy-webpack-plugin": "^4.2.0",
+    "path": "^0.12.7",
+    "uglifyjs-webpack-plugin": "^1.0.0",
+    "webpack": "^3.8.1",
+    "webpack-merge": "^4.1.0"
+  },
+  "dependencies": {
+    "angular": "^1.5.0-rc.0"
+  }
+}
