+++ src/app/components/index.js
@@ -1,0 +1,5 @@
+
+export { exportController } from './export/exportController'
+export { importController } from './import/importController'
+export { inventoryController } from './inventory/inventoryController'
+export { searchController } from './search/searchController'
