+++ webpack.dev.js
@@ -1,0 +1,8 @@
+
+const merge = require('webpack-merge');
+const common = require('./webpack.common.js');
+
+module.exports = merge(common, {
+  devtool: 'inline-source-map',
+  stats: 'detailed'
+})
